src/AuthFlow/AdminUserFlow/api/v1/app/requirements.txt
bcrypt
Flask
Flask-Bcrypt
Flask-SQLAlchemy
importlib-metadata
SQLAlchemy
python-decouple
psycopg2-binary
python-dotenv